+++ flexiapi/resources/views/admin/account/provisioning_email/create.blade.php
@@ -3,13 +3,13 @@
 @section('breadcrumb')
     @include('admin.account.parts.breadcrumb_accounts_index')
     @include('admin.account.parts.breadcrumb_accounts_show', ['account' => $account])
-    <li class="breadcrumb-item active" aria-current="page">{{ __('Reset password') }}</li>
+    <li class="breadcrumb-item active" aria-current="page">{{ __('Provisioning') }}</li>
 @endsection
 
 @section('content')
 
 <header>
-    <h1><i class="ph ph-envelope"></i> {{ __('Reset password') }}</h1>
+    <h1><i class="ph ph-envelope"></i> {{ __('Provisioning') }}</h1>
 </header>
 
 <p>{{ __('An email will be sent to :email with a QR Code and provisioning link.', ['email' => $account->email]) }}</p>